VZ-9610 remove unused code (#174)
